Propagate isDisabled property through boxed elements and codelists as select
With this change a user can define a boxed element as disabled. This commit propagates
that parameter through these templates to render the fields as readonly.

For example, you could use it to set the /mdb:MD_Metadata/mdb:identificationInfo/mri:MD_DataIdentification/mri:citation/cit:CI_Citation/cit:citedResponsibleParty
children fields as not editable because you want the users to select the
contact from a search contact widget.